[POL-17] Доделал работу с новой базой данных

+++ app/src/main/java/com/example/polycareer/data/database/model/UsersAnswersEntity.kt
@@ -22,5 +22,5 @@ import androidx.room.PrimaryKey
 data class UsersAnswersEntity(
     @ColumnInfo(name = "user_id") val userId: Long,
     @ColumnInfo(name = "answer_id") val answerId: Long,
-    @ColumnInfo(name = "try_number") val tryNumber: Int = 0
+    @ColumnInfo(name = "try_number") val tryNumber: Long = 0
 )
